[tests] Restore individual NUnit interface collection cases
Move the six behavioral cases into the existing Mono.Android.NET-Tests project with NUnit assertions and individual discovery. Exercise managed dictionary enumeration instead of raw Map.Entry traversal, and share the unbound Java fixture and JNI holder with a minimal NativeAOT rooting probe.

Keep the exclusive NativeAOT dependency-graph assertions in MSBuildDeviceIntegration without a custom six-case runner or pass counter.
