--- tests/MSBuildDeviceIntegration/Resources/InterfaceCollectionApp/ExtendedValueProvider.java
@@ -1,5 +1,0 @@
-package net.dot.android.test;
-
-public interface ExtendedValueProvider extends ValueProvider {
-	int getOtherValue();
-}
