--- tests/MSBuildDeviceIntegration/Resources/InterfaceCollectionApp/ValueProvider.java
@@ -1,5 +1,0 @@
-package net.dot.android.test;
-
-public interface ValueProvider {
-	int getValue();
-}
