+++ tests/Mono.Android-Tests/Mono.Android-Tests/InterfaceCollection.proguard.cfg
